improvements to the layout, update to 'my jams' popup.
