+++ functions/index.js
@@ -10,6 +10,8 @@ const processJam = jam => ({
     name: jam.name,
     dateUpdated: jam.dateUpdated || jam.dateCreated,
     runtime: jam.runtime,
+    language: jam.language || 'javascript',
+    dialect: jam.dialect || 'en',
     features: {
         lineCount: jam.features.map(i => i.source.split('\n').length).reduce((a, b) => a + b),
         scenarioCount: jam.features.map(i => i.source.split('\n').filter(l => l.trim().indexOf('Scenario') === 0).length).reduce((a, b) => a + b)
@@ -26,6 +28,7 @@ exports.getMyJams = functions.https.onCall((data, context) => {
     }
     const limit = parseInt(data.limit, 10) || 10;
     const after = data.after;
+    console.log(`After: ${after}`);
 
     let query = firestore.collection('jams')
         .where('uid', '==', uid)
