+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "testjam-io",
-  "version": "0.8.2",
+  "version": "0.8.3",
   "description": "React web application for testjam.io",
   "main": "index.js",
   "scripts": {
@@ -39,6 +39,7 @@
     "firebase": "^9.8.4",
     "gherkin": "^9.0.0",
     "indent.js": "^0.3.5",
+    "javascript-time-ago": "^2.5.6",
     "lodash": "^4.17.15",
     "prop-types": "^15.8.1",
     "react": "^16.2.0",
@@ -46,6 +47,7 @@
     "react-dom": "^16.2.0",
     "react-dropdown": "^1.7.0",
     "react-notifications": "^1.6.0",
+    "react-paginate": "^8.1.3",
     "react-resize-detector": "^7.1.2",
     "react-router-bootstrap": "^0.24.4",
     "react-router-dom": "^4.2.2",
@@ -56,8 +58,8 @@
     "styled-components": "^5.3.5"
   },
   "build": {
-    "version": "0.8.2",
-    "date": "2022-07-06T17:08:00Z"
+    "version": "0.8.3",
+    "date": "2022-07-07T00:54:28.980Z"
   },
   "devDependencies": {
     "browserify": "^17.0.0",
