+++ src/views/HeaderView.jsx
@@ -8,11 +8,11 @@ import { NotificationManager } from 'react-notifications';
 import Dropdown from 'react-dropdown';
 import Popup from 'reactjs-popup';
 
+import AboutSite from '../components/menuItems/AboutSitePopup';
+import RunTestsButton from '../components/buttons/RunTestsButton';
+import SignInButton from '../components/buttons/SignInButton';
+
 import app from '../app';
-
-const { build } = require('../../package.json');
-
-const setImmediate = (fn) => setTimeout(fn, 0);
 
 export default class HeaderView extends React.Component {
     constructor(props) {
@@ -21,10 +21,7 @@ export default class HeaderView extends React.Component {
         this.editName = this.editName.bind(this);
         this.setName = this.setName.bind(this);
         this.setTheme = this.setTheme.bind(this);
-        this.runTest = this.runTest.bind(this);
         this.toggleLike = this.toggleLike.bind(this);
-        this.signIn = this.signIn.bind(this);
-        this.signOut = this.signOut.bind(this);
         this.save = this.save.bind(this);
         this.fork = this.fork.bind(this);
         this.copyLink = this.copyLink.bind(this);
@@ -76,6 +73,7 @@ export default class HeaderView extends React.Component {
         app.on('nameChanged', (name) => this.setState({ name, isEditName: false }));
 
         app.on('myJamsLoaded', (myJams) => this.setState({ myJams, isMyJamsLoading: false }));
+        app.on('jamsAfterChanged', (after) => this.setState({ after }));
     }
 
     isOwner() {
@@ -106,10 +104,6 @@ export default class HeaderView extends React.Component {
         app.setName(e.target.value);
     }
 
-    runTest() {
-        app.run();
-    }
-
     toggleLike() {
         const { likeChanging, isLiked } = this.state;
         if (!likeChanging) {
@@ -119,14 +113,6 @@ export default class HeaderView extends React.Component {
                 app.like();
             }
         }
-    }
-
-    signIn() {
-        app.signIn();
-    }
-
-    signOut() {
-        app.signOut();
     }
 
     save() {
@@ -143,39 +129,6 @@ export default class HeaderView extends React.Component {
         if (!isForking && allowFork) {
             app.fork();
         }
-    }
-
-    loadMyJams() {
-        app.getMyJams();
-    }
-
-    renderTestButton() {
-        let innerButton = (
-            <span>
-                <i className="fa fa-play-circle" />
-&nbsp;&nbsp;Run Tests
-            </span>
-        );
-        let buttonClass = 'btn btn-run';
-        const { isTestRunning, isTestEnabled } = this.state;
-        if (isTestRunning) {
-            innerButton = (
-                <span>
-                    <i className="fa fa-spin fa-circle-o-notch" />
-&nbsp;Running...
-                </span>
-            );
-            buttonClass = 'btn btn-running';
-        } else if (!isTestEnabled) {
-            innerButton = (
-                <span>
-                    <i className="fa fa-ban" />
-&nbsp;&nbsp;Run Tests
-                </span>
-            );
-            buttonClass = 'btn btn-norun';
-        }
-        return <div className={buttonClass} role="button" tabIndex={0} onClick={this.runTest} onKeyDown={app.handleEnter(this.runTest)}>{innerButton}</div>;
     }
 
     renderLikeButton() {
@@ -260,170 +213,15 @@ export default class HeaderView extends React.Component {
         return '';
     }
 
-    renderUserDisplay() {
-        const { user } = this.state;
-        return (
-            <div className="btn user-info">
-                <img src={user.photo} alt="profile" />
-                <span className="user-name">{user.name}</span>
-            </div>
-        );
-    }
-
-    renderMyJams() {
-        const { user, isMyJamsLoading, myJams } = this.state;
-        return (
-            <Popup
-                trigger={(isOpen) => {
-                    if (isOpen) { setImmediate(() => this.loadMyJams()); }
-                    return (
-                        <div className="Dropdown-option">
-                            <i className="fa fa-book" />
-&nbsp;&nbsp;My Jams&nbsp;(
-                            {user.jamsCount || 0}
-                            )
-                        </div>
-                    );
-                }}
-                modal
-                contentStyle={{ padding: '10px 20px 10px 20px' }}
-                closeOnDocumentClick
-            >
-                <div style={{ width: '100%' }}>
-                    {isMyJamsLoading
-                        ? <div className="popup-myjams-loading"><i className="fa fa-spin fa-circle-o-notch" /></div>
-                        : (
-                            <div className="popup-myjams">
-                                {myJams.map((jam) => (
-                                    <a className="myjam" href={`${window.location.origin}/?p=${jam.id}`} key={jam.id}>
-                                        <h3>{jam.name}</h3>
-                                        <div>
-                                            Last Updated
-                                            {new Date(jam.dateUpdated).toLocaleString()}
-                                        </div>
-                                        <div>
-                                            Runtime:
-                                            {jam.runtime}
-                                        </div>
-                                        <div>
-                                            Scenarios:
-                                            {jam.features.scenarioCount}
-                                            {' '}
-                                            (
-                                            {jam.features.lineCount}
-                                            {' '}
-                                            lines)
-                                        </div>
-                                        <div>
-                                            Step Definitions: JavaScript (
-                                            {jam.stepDefinitions.lineCount}
-                                            {' '}
-                                            lines)
-                                        </div>
-                                    </a>
-                                ))}
-                            </div>
-                        )}
-                </div>
-            </Popup>
-        );
-    }
-
-    renderSignInArea() {
-        const { user } = this.state;
-        if (user.uid) {
-            return (
-                <Popup
-                    trigger={this.renderUserDisplay()}
-                    position="bottom right"
-                    closeOnDocumentClick
-                    closeOnEscape
-                    mouseLeaveDelay={300}
-                    contentStyle={{ padding: '0px', border: 'none' }}
-                    arrow={false}
-                >
-                    <div className="Dropdown-menu" aria-expanded="true">
-                        {this.renderMyJams()}
-                        <div className="Dropdown-option" onClick={this.signOut} onKeyDown={app.handleEnter(this.signOut)} role="button" tabIndex={0}>
-                            <i className="fa fa-sign-out" />
-&nbsp;&nbsp;Sign Out
-                        </div>
-                    </div>
-                </Popup>
-            );
-        }
-        return <div onClick={this.signIn} onKeyDown={app.handleEnter(this.signIn)} className="btn user-signin" role="button" tabIndex={0}>Log In/Sign Up</div>;
-    }
-
     renderMenu() {
+        const {
+            isTestRunning, isTestEnabled, user, isMyJamsLoading, myJams, after,
+        } = this.state;
         return (
             <div className="menu">
-                {this.renderTestButton()}
-                {this.renderSignInArea()}
+                <RunTestsButton isTestRunning={isTestRunning} isTestEnabled={isTestEnabled} />
+                <SignInButton user={user} isMyJamsLoading={isMyJamsLoading} myJams={myJams} after={after} />
             </div>
-        );
-    }
-
-    renderAboutSite() {
-        return (
-            <Popup
-                trigger={<div className="Dropdown-option">About testjam.io</div>}
-                modal
-                contentStyle={{ padding: '10px 20px 10px 20px' }}
-                closeOnDocumentClick
-            >
-                <div style={{ width: '100%' }}>
-                    <h2 style={{ marginBottom: '0' }}>
-                        testjam.io version
-                        {build.version}
-                    </h2>
-                    <p>
-                        Last Updated on
-                        {new Date(build.date).toDateString()}
-                    </p>
-                    <br />
-                    <h3>What is TestJam?</h3>
-                    <p>
-                        Welcome to testjam.io, an online coding environment specifically designed for Cucumber and Gherkin. When I first started building this project, I realized that most major programming languages and frameworks have some sort of online sandbox environment for sharing code examples, but Cucumber did not. As BDD gets more popular in the modern QA toolkit, the need to quickly share and experiment with testing tools over the Internet will increase.
-                    </p>
-                    <h3>Project Objectives</h3>
-                    <ul>
-                        <li>Deliver an easy-to-use web-based tool for creating and sharing feature files and their accompanying step definitions.</li>
-                        <li>Improve the quality of online documentation and Q&amp;A responses regarding Cucumber and Gherkin.</li>
-                        <li>Provide a useful service to the wider QA automation community.</li>
-                        <li>
-                            <i>Personal Goal:</i>
-&nbsp;Avoid boredom during the COVID-19 quarantine by developing and launching a new web-based software tool.
-                        </li>
-                    </ul>
-                    <h3>About Me</h3>
-                    <p>
-                        My name is Steven, and I&apos;m an automation engineer and software developer. I have been using Cucumber (
-                        <a href="https://specflow.org/" target="_blank" rel="noreferrer">SpecFlow</a>
-                        {' '}
-                        and
-                        {' '}
-                        <a href="https://github.com/cucumber/cucumber-js" target="_blank" rel="noreferrer">CucumberJS</a>
-                        ) at various organizations since 2014, both as a developer and an automation engineer. I hope to continue to learn more about automated testing in the future and also help mentor others on software and maintaining quality during a project.
-                        <br />
-                        <br />
-                    </p>
-                    <a className="link-item" href="https://www.github.com/stevenmhunt" target="_blank" rel="noreferrer">
-                        <i className="fa fa-github" />
-&nbsp;@stevenmhunt
-                    </a>
-                    <a className="link-item" href="https://www.twitter.com/stevenmhunt" target="_blank" rel="noreferrer">
-                        <i className="fa fa-twitter" />
-&nbsp;@stevenmhunt
-                    </a>
-                    <a className="link-item" href="http://www.linkedin.com/in/stevenmhunt" target="_blank" rel="noreferrer">
-                        <i className="fa fa-linkedin" />
-&nbsp;stevenmhunt
-                    </a>
-                    <br />
-                    <br />
-                </div>
-            </Popup>
         );
     }
 
@@ -459,7 +257,7 @@ export default class HeaderView extends React.Component {
                                 <Dropdown options={this.themes} value={theme} onChange={this.setTheme} />
                             </div>
                             <a className="Dropdown-option" href={`${window.location.origin}/?p=0Djdc5ps8TE1ssv558rU`}>Getting Started</a>
-                            {this.renderAboutSite()}
+                            <AboutSite />
                             <a className="Dropdown-option" href="https://cucumber.io/docs/gherkin/" target="_blank" rel="noreferrer">
                                 <i className="fa fa-external-link" />
 &nbsp;&nbsp;Learn more about Gherkin
@@ -491,6 +289,7 @@ export default class HeaderView extends React.Component {
                                 <div className="fork-tooltip">
                                     <div>
                                         Forked from
+                                        {' '}
                                         <b>{fork.name}</b>
 &nbsp;&nbsp;
                                         <a href={`${window.location.origin}/?p=${fork.id}`} target="_blank" rel="noreferrer">
